Use Decimal for certain OpenAPI params to avoid losing precision

+++ fastapi/openapi/models.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from enum import Enum
 from typing import Any, Callable, Dict, Iterable, List, Optional, Union
 
@@ -82,11 +83,11 @@ class ExternalDocumentation(BaseModel):
 class SchemaBase(BaseModel):
     ref: Optional[str] = Field(None, alias="$ref")
     title: Optional[str] = None
-    multipleOf: Optional[float] = None
-    maximum: Optional[float] = None
-    exclusiveMaximum: Optional[float] = None
-    minimum: Optional[float] = None
-    exclusiveMinimum: Optional[float] = None
+    multipleOf: Optional[Decimal] = None
+    maximum: Optional[Decimal] = None
+    exclusiveMaximum: Optional[Decimal] = None
+    minimum: Optional[Decimal] = None
+    exclusiveMinimum: Optional[Decimal] = None
     maxLength: Optional[int] = Field(None, gte=0)
     minLength: Optional[int] = Field(None, gte=0)
     pattern: Optional[str] = None
